Add EnvReader utility class
A new utility class called EnvReader has been added. This class includes a static method to read an environment variable. This change will allow us to access environment variables in a more modular and maintainable way. If the environment variable is not found, an appropriate error message will be logged.