Add 'appName' to trace logs and associated files
This commit involves adding an 'appName' property to the existing HTTP trace logs. By doing so, it provides more detailed logging information, which will improve the debugging process. The 'appName' is extracted from a MDC property map and is added to the 'log' object in 'JsonEncoder', and 'HttpTraceDTO' to be used wherever necessary.