Add Maven dependency and application information in logs
Added 'maven-model' dependency to pom.xml and modified log generation to include application name and version automatically. Also, implemented mechanism to generate UUID, if not received in the request header, to be used as request ID. These changes will enhance the traceability and debugging process by providing comprehensive and useful information in log data about the application and every request it handles.