Import MDC and adjust correlationID management in CustomHttpTraceFilter
Adjusted the management of correlationId in CustomHttpTraceFilter class. The org.slf4j.MDC is now imported and used to remove the CORRELATION_ID_LOG_VAR_NAME once the logging operation is complete. This ensures a better cleanup of the Mapped Diagnostic Context and prevents potential misuse in later phases of the request/response lifecycle.