Refactor requestId retrieval in RequestInfo class
Changed the sequence of getting requestId in RequestInfo class. Previously, it first attempted to get the requestId from MDC, and if not found, then from the CORRELATION_ID_HEADER_NAME. However, now it first attempts to get the requestId from the CORRELATION_ID_HEADER_NAME, and only if not found, it tries to get it from MDC. This ensures a more logical order of retrieval.